Use C-style strings.
As of https://github.com/ros2/rclcpp/pull/1442 , the RCLCPP_
macros no longer accept std::string arguments, and only C-style
strings.

https://index.ros.org/doc/ros2/Releases/Release-Galactic-Geochelone/#change-in-rclcpp-s-logging-macros
has more information about it.

This PR changes all of the use of std::string to C-style strings,
which will allow this package to build against Rolling (and
hence Galactic).

Note that this change is entirely backwards compatible, so it
can be used for both Foxy and Rolling.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>